refactor(MediaType): ContentType -> MediaType으로 이름 변경
